third_party/blink/web_tests/external/wpt/html/dom/aria-attribute-reflection-expected.txt

This is a testharness.js-based test.
[FAIL] aria-colindextext attribute reflects.
  assert_equals: expected (string) "x" but got (undefined) undefined
[FAIL] aria-rowindextext attribute reflects.
  assert_equals: expected (string) "x" but got (undefined) undefined
Harness: the test ran to completion.
